rename DesignableBannerV2 to DesignableBanner across the board, duplicate import in ReaderRevenueBanner till fixed in SDC
