+++ dotcom-rendering/src/components/StickyBottomBanner/ReaderRevenueBanner.tsx
@@ -312,9 +312,9 @@ export const ReaderRevenueBanner = ({
 			  import(`../marketing/banners/signInPrompt/SignInPromptBanner`)
 			: name === 'DesignableBannerV2'
 			? /* webpackChunkName: "designable-banner-v2" */
-			  import(`../marketing/banners/designableBanner/DesignableBannerV2`)
+			  import(`../marketing/banners/designableBanner/DesignableBanner`)
 			: /* webpackChunkName: "designable-banner" */
-			  import(`../marketing/banners/designableBanner/DesignableBannerV2`)
+			  import(`../marketing/banners/designableBanner/DesignableBanner`)
 		)
 			.then((bannerModule: { [key: string]: React.ElementType }) => {
 				setBanner(() => bannerModule[name] ?? null);
