+++ dotcom-rendering/src/components/marketing/banners/designableBanner/DesignableBanner.tsx
